# Improve bind performance