@ffenix113 2020-01-04
After new simplification:

name                    old time/op    new time/op    delta
BindbindData-8            21.2µs ± 2%    11.7µs ± 1%  -44.80%  (p=0.000 n=16+19)
BindbindDataWithTags-8    22.1µs ± 1%    14.2µs ± 2%  -35.63%  (p=0.000 n=20+18)

name                    old alloc/op   new alloc/op   delta
BindbindData-8            2.40kB ± 0%    0.80kB ± 0%  -66.65%  (p=0.000 n=20+20)
BindbindDataWithTags-8    2.31kB ± 0%    0.98kB ± 0%  -57.42%  (p=0.000 n=20+20)

name                    old allocs/op  new allocs/op  delta
BindbindData-8               297 ± 0%        49 ± 0%  -83.50%  (p=0.000 n=20+20)
BindbindDataWithTags-8       267 ± 0%        51 ± 0%  -80.90%  (p=0.000 n=20+20)